+++ src/main/java/org/openlmis/fulfillment/AuditLogInitializer.java
@@ -19,12 +19,13 @@ import static org.openlmis.fulfillment.util.Pagination.DEFAULT_PAGE_NUMBER;
 
 import java.util.List;
 import java.util.Map;
-import javax.annotation.Resource;
 import org.javers.core.Javers;
 import org.javers.core.metamodel.object.CdoSnapshot;
 import org.javers.repository.jql.QueryBuilder;
 import org.javers.spring.annotation.JaversSpringDataAuditable;
 import org.openlmis.fulfillment.domain.BaseEntity;
+import org.openlmis.fulfillment.i18n.MessageKeys;
+import org.openlmis.fulfillment.repository.BaseAuditableRepository;
 import org.slf4j.ext.XLogger;
 import org.slf4j.ext.XLoggerFactory;
 import org.slf4j.profiler.Profiler;
@@ -36,10 +37,7 @@ import org.springframework.core.annotation.Order;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
 
 /**
  * AuditLogInitializer runs after its associated Spring application has loaded.
@@ -50,15 +48,17 @@ import org.springframework.transaction.annotation.Transactional;
 @Component
 @Profile("refresh-db")
 @Order(20)
-@Transactional
 public class AuditLogInitializer implements CommandLineRunner {
   private static final XLogger LOGGER = XLoggerFactory.getXLogger(AuditLogInitializer.class);
 
-  @Autowired
   private ApplicationContext applicationContext;
-
-  @Resource(name = "javersProvider")
   private Javers javers;
+
+  @Autowired
+  public AuditLogInitializer(ApplicationContext applicationContext, Javers javers) {
+    this.applicationContext = applicationContext;
+    this.javers = javers;
+  }
 
   /**
    * This method is part of CommandLineRunner and is called automatically by Spring.
@@ -88,18 +88,19 @@ public class AuditLogInitializer implements CommandLineRunner {
   }
 
   private void createSnapshots(Object bean) {
-    if (bean instanceof PagingAndSortingRepository) {
-      createSnapshots((PagingAndSortingRepository<?, ?>) bean);
-    } else if (bean instanceof CrudRepository) {
-      createSnapshots((CrudRepository<?, ?>) bean);
+    if (bean instanceof BaseAuditableRepository) {
+      createSnapshots((BaseAuditableRepository<?, ?>) bean);
+    } else {
+      LOGGER.warn("The repository should implement findAllWithoutSnapshots method"
+          + "from BaseAuditableRepository with appropriate query");
     }
   }
 
-  private void createSnapshots(PagingAndSortingRepository<?, ?> repository) {
+  private void createSnapshots(BaseAuditableRepository<?, ?> repository) {
     Pageable pageable = new PageRequest(DEFAULT_PAGE_NUMBER, 2000);
 
     while (true) {
-      Page<?> page = repository.findAll(pageable);
+      Page<?> page = repository.findAllWithoutSnapshots(pageable);
 
       if (!page.hasContent()) {
         break;
@@ -109,11 +110,6 @@ public class AuditLogInitializer implements CommandLineRunner {
 
       pageable = pageable.next();
     }
-  }
-
-  private void createSnapshots(CrudRepository<?, ?> repository) {
-    //... retrieve all of its domain objects and...
-    repository.findAll().forEach(this::createSnapshot);
   }
 
   private void createSnapshot(Object object) {
@@ -126,8 +122,11 @@ public class AuditLogInitializer implements CommandLineRunner {
     List<CdoSnapshot> snapshots = javers.findSnapshots(jqlQuery.build());
 
     //If there are no snapshots of the domain object, then take one
-    if (snapshots.size() == 0) {
+    if (snapshots.isEmpty()) {
       javers.commit("System: AuditLogInitializer", baseEntity);
+    } else {
+      LOGGER.info(MessageKeys.ERROR_JAVERS_EXISTING_ENTRY,
+          baseEntity.getClass(), baseEntity.getId());
     }
   }