+++ src/main/java/org/openlmis/fulfillment/i18n/MessageKeys.java
@@ -46,6 +46,7 @@ public abstract class MessageKeys {
   private static final String PARAMETER = "parameter";
   private static final String TEMPLATE = "template";
   private static final String LINE_ITEMS = "lineItems";
+  private static final String JAVERS = "javers";
 
   private static final String REQUIRED = "required";
   private static final String NOT_FOUND = "notFound";
@@ -177,6 +178,9 @@ public abstract class MessageKeys {
 
   public static final String EVENT_MISSING_SOURCE_DESTINATION =
       join(ERROR_PREFIX, STOCK_EVENT, "missingSourceDestination");
+
+  public static final String ERROR_JAVERS_EXISTING_ENTRY =
+      join(ERROR_PREFIX, JAVERS, "entryAlreadyExists");
 
   private static String join(String... params) {
     return String.join(DELIMITER, Arrays.asList(params));